fix: the type of param

+++ src/ValueObjects/Fine.php
@@ -10,7 +10,7 @@ use AsaasPhpSdk\ValueObjects\Base\AbstractStructuredValueObject;
 final class Fine extends AbstractStructuredValueObject
 {
 	public function __construct(
-		public readonly int $value,
+		public readonly float $value,
 		public readonly FineType $type,
 	) {}
